Simlify encryption config for aot (#2173)
