+++ spring-cloud-config-server/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,5 @@
 org.springframework.cloud.config.server.config.ConfigServerAutoConfiguration
+org.springframework.cloud.config.server.config.RsaEncryptionAutoConfiguration
+org.springframework.cloud.config.server.config.DefaultTextEncryptionAutoConfiguration
 org.springframework.cloud.config.server.config.EncryptionAutoConfiguration
 org.springframework.cloud.config.server.config.VaultEncryptionAutoConfiguration
